adding of SevenStonesInARow.java six die, eight live heuristic doesn't check corners yet, ie for rows of 5 in the corner so 6th move would keep the group alive

+++ src/main/Board.java
@@ -1,5 +1,6 @@
 package main;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 
 public class Board {
@@ -48,7 +49,7 @@ public class Board {
     }
 
     // Get int representation of board position
-    public int get(int x, int y) {
+    public byte get(int x, int y) {
         return boardRep[x][y];
     }
